requirements.txt
langchain==0.0.350
langchain-community==0.0.16
langchain-core==0.1.20
langchain-google-genai==0.0.6
google-generativeai==0.3.2
faiss-cpu==1.11.0
streamlit==1.31.0
numpy==1.26.3
python-dotenv==1.0.0
tiktoken==0.5.2
sentence-transformers==2.5.0
langchain-text-splitters==0.0.1
